big renderWithRedux refactor
