+++ test/components/grouping_idea_card.test.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { render } from "@testing-library/react"
 import "@testing-library/jest-dom"
-import renderWithRedux from "../support/js/render_with_redux"
+import { renderWithRedux } from "../support/js/jest_test_helper"
 
 import { GroupingIdeaCard } from "../../web/static/js/components/grouping_idea_card"
 import { COLLISION_BUFFER } from "../../web/static/js/services/collisions"
